refactor(app): extract chat opening
